Added PlayerQueryDSL, refactored PlayerRepository
