+++ src/main/java/com/ardaslegends/repository/player/PlayerRepositoryCustom.java
@@ -1,0 +1,10 @@
+package com.ardaslegends.repository.player;
+
+import com.ardaslegends.domain.Player;
+
+import java.util.Set;
+
+public interface PlayerRepositoryCustom {
+    Player queryByDiscordId(String discordId);
+    Set<Player> queryByDiscordId(String[] discordIds);
+}
