+++ src/main/java/com/ardaslegends/service/FactionService.java
@@ -3,7 +3,7 @@ package com.ardaslegends.service;
 import com.ardaslegends.domain.Faction;
 import com.ardaslegends.domain.Player;
 import com.ardaslegends.repository.FactionRepository;
-import com.ardaslegends.repository.PlayerRepository;
+import com.ardaslegends.repository.player.PlayerRepository;
 import com.ardaslegends.service.dto.UpdateFactionLeaderDto;
 import com.ardaslegends.service.dto.faction.UpdateStockpileDto;
 import com.ardaslegends.service.exceptions.FactionServiceException;
