+++ src/main/java/com/ardaslegends/service/war/WarService.java
@@ -4,7 +4,7 @@ import com.ardaslegends.domain.Faction;
 import com.ardaslegends.domain.Player;
 import com.ardaslegends.domain.war.War;
 import com.ardaslegends.repository.FactionRepository;
-import com.ardaslegends.repository.PlayerRepository;
+import com.ardaslegends.repository.player.PlayerRepository;
 import com.ardaslegends.repository.WarRepository;
 import com.ardaslegends.service.AbstractService;
 import com.ardaslegends.service.dto.war.CreateWarDto;
