+++ src/test/java/com/ardaslegends/repository/PlayerRepositoryTest.java
@@ -1,6 +1,7 @@
 package com.ardaslegends.repository;
 
 import com.ardaslegends.domain.*;
+import com.ardaslegends.repository.player.PlayerRepository;
 import lombok.val;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
